Refactor code base
- Migrate components of __main__ into separate modules
- Fix pre-commit action
