--- robokitty/_constants.py
@@ -1,3 +1,0 @@
-# Serial config
-DEFAULT_PORT = "/dev/ttyUSB0"
-DEFAULT_BAUD = 1000000
