+++ robokitty/_robokitty/__init__.py
@@ -1,0 +1,11 @@
+from .walker import QuadrupedWalker
+from .utils import print_diagnostics, run_keyboard_control
+from .constants import DEFAULT_BAUD, DEFAULT_PORT
+
+__all__ = [
+    "QuadrupedWalker",
+    "print_diagnostics",
+    "run_keyboard_control",
+    "DEFAULT_BAUD",
+    "DEFAULT_PORT",
+]
